21-02-2025 uploaded

+++ Daily Assignments/20-02-2025/Rest-API-Methods-Using-Mongodb/src/main/java/com/example/dao/Repository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.example.model.EmpDetails;
 
-public interface Repository extends MongoRepository<EmpDetails,Long>{
+public interface Repository extends MongoRepository<EmpDetails,Integer>{
 }
